IR GFSM

+++ src/main/java/org/biiig/dmgm/impl/algorithms/subgraph_mining/csm/Interestingness.java
@@ -2,11 +2,11 @@ package org.biiig.dmgm.impl.algorithms.subgraph_mining.csm;
 
 import java.util.function.BiFunction;
 
-public interface Interestingness extends BiFunction<Integer, Integer, Boolean> {
-  boolean isInteresting(int categoryCount, int totalCount);
+public interface Interestingness extends BiFunction<Float, Float, Boolean> {
+  boolean isInteresting(Float categoryCount, Float totalCount);
 
   @Override
-  default Boolean apply(Integer categoryCount, Integer totalCount) {
+  default Boolean apply(Float categoryCount, Float totalCount) {
     return isInteresting(categoryCount, totalCount);
   }
 }
